Ds/saturday updates

+++ client/store/cart.js
@@ -1,13 +1,14 @@
-import axios from "axios";
+import axios from 'axios';
 
-const TOKEN = "token";
+const TOKEN = 'token';
 
 const initialState = {};
 
 // action types
-const SET_CART = "SET_CART";
-const DELETE_ITEM = "DELETE_ITEM";
-const ADDED_ITEM = "ADDED_ITEM";
+const SET_CART = 'SET_CART';
+const DELETE_ITEM = 'DELETE_ITEM';
+const ADDED_ITEM = 'ADDED_ITEM';
+const SUBTRACTED_ITEM = 'SUBTRACTED_ITEM';
 
 //action creators
 export const setCart = (cartObj) => {
@@ -17,10 +18,10 @@ export const setCart = (cartObj) => {
   };
 };
 
-export const deleteItem = (cart) => {
+export const deleteItem = (productId) => {
   return {
     type: DELETE_ITEM,
-    cart,
+    productId,
   };
 };
 
@@ -28,6 +29,13 @@ export const addedItem = (item) => {
   return {
     type: ADDED_ITEM,
     item,
+  };
+};
+
+export const subtractedItem = (productId) => {
+  return {
+    type: SUBTRACTED_ITEM,
+    productId,
   };
 };
 
@@ -49,7 +57,8 @@ export const fetchCart = (userId) => {
 };
 
 export const addToCart = (userId, productId) => {
-  return async (dispatch) => {
+  return async (dispatch, getState) => {
+    const cart = getState().cart;
     const token = window.localStorage.getItem(TOKEN);
     try {
       if (token) {
@@ -58,16 +67,25 @@ export const addToCart = (userId, productId) => {
             authorization: token,
           },
         };
-        const { data: product } = await axios.post(
-          `/api/cart/${userId}/${productId}`,
-          null,
-          headers
-        );
-        console.log("Logged in: ", product);
-        dispatch(addedItem(product));
+        if (cart[productId]) {
+          const { data: product } = await axios.put(
+            `/api/cart/${userId}/${productId}`,
+            { increment: 1 },
+            {
+              headers: { authorization: token },
+            }
+          );
+          dispatch(addedItem(product));
+        } else {
+          const { data: product } = await axios.post(
+            `/api/cart/${userId}/${productId}`,
+            null,
+            headers
+          );
+          dispatch(addedItem(product));
+        }
       } else {
         const { data: product } = await axios.get(`/api/products/${productId}`);
-        console.log("Not Logged in:", product);
         dispatch(addedItem(product));
       }
     } catch (err) {
@@ -80,9 +98,11 @@ export const deleteFromCart = (userId, productId) => {
   return async (dispatch) => {
     try {
       const token = window.localStorage.getItem(TOKEN);
-      await axios.delete(`/api/cart/${userId}/${productId}`, {
-        headers: { authorization: token },
-      });
+      if (token) {
+        await axios.delete(`/api/cart/${userId}/${productId}`, {
+          headers: { authorization: token },
+        });
+      }
       dispatch(deleteItem(productId));
     } catch (err) {
       throw err;
@@ -90,14 +110,37 @@ export const deleteFromCart = (userId, productId) => {
   };
 };
 
-//reducer
+// remove cart thunk make check quantity
+// if quantity is 1 dispatch delete from cart
+// else axios put call to update quantity
+export const subtractFromCart = (userId, productId) => {
+  return async (dispatch, getState) => {
+    try {
+      const cart = getState().cart;
+      if (cart[productId].quantity === 1) {
+        dispatch(deleteFromCart(userId, productId));
+      } else {
+        await axios.put(
+          `/api/cart/${userId}/${productId}`,
+          { decrement: 1 },
+          {
+            headers: { authorization: token },
+          }
+        );
+        dispatch(subtractedAdded(productId));
+      }
+    } catch (error) {
+      throw error;
+    }
+  };
+};
+
 export default function cartsReducer(state = initialState, action) {
+  let newCart = { ...state };
   switch (action.type) {
     case SET_CART:
       return action.cartObj;
     case ADDED_ITEM: {
-      console.log(action);
-      let newCart = { ...state };
       if (newCart[action.item.id]) {
         newCart[action.item.id].quantity++;
       } else {
@@ -112,9 +155,11 @@ export default function cartsReducer(state = initialState, action) {
       return newCart;
     }
     case DELETE_ITEM: {
-      let newCart = { ...state };
-      delete newCart[action.cart];
+      delete newCart[action.productId];
       return newCart;
+    }
+    case SUBTRACTED_ITEM: {
+      newCart[action.productId].quantity--;
     }
     default:
       return state;
